added permissions check
