+++ packages/osc-api/utils/types.ts
@@ -1,0 +1,5 @@
+export interface PermissionsProps {
+    readonly [key: string]: string[];
+}
+
+export type UserPermissionsFn = (userId: number) => Promise<PermissionsProps>;
